requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
djangorestframework
djangorestframework-simplejwt
django-allauth
dj-rest-auth
requests
cryptography
python-decouple
